- Fixed duplicate page_count method in document.rs
- Resolved conflicts in batch processing examples
- Fixed streaming support tests imports
- Updated page streamer constructors
- Resolved worker.rs execution flow
- Fixed error.rs comment formatting
- Updated lazy_loader object insertions
- Fixed recovery module page dimensions
- Resolved repair.rs return statements
- Fixed scanner.rs object detection test data
- Updated writer.rs date formatting to UTC
- Fixed merge_tests.rs assertions
- Merged CLAUDE.md sessions and CI/CD guidelines
- Fixed batch_processing_tests for floating-point comparisons

All tests passing (1206 tests)

+++ oxidize-pdf-core/src/recovery/mod.rs
@@ -232,7 +232,7 @@ impl PdfRecovery {
         for i in 0..page_count {
             if let Ok(page) = pdf_doc.get_page(i) {
                 // Simple conversion - would need proper implementation
-                let new_page = crate::Page::new(page.width() as f64, page.height() as f64);
+                let new_page = crate::Page::new(page.width(), page.height());
                 doc.add_page(new_page);
             }
         }
